Reduce code duplication in various batch systems
