+++ src/toil/batchSystems/__init__.py
@@ -11,4 +11,37 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 from __future__ import absolute_import
+
+
+class MemoryString:
+    def __init__(self, string):
+        if string[-1] == 'K' or string[-1] == 'M' or string[-1] == 'G' or string[-1] == 'T':
+            self.unit = string[-1]
+            self.val = float(string[:-1])
+        else:
+            self.unit = 'B'
+            self.val = float(string)
+        self.bytes = self.byteVal()
+
+    def __str__(self):
+        if self.unit != 'B':
+            return str(self.val) + self.unit
+        else:
+            return str(self.val)
+
+    def byteVal(self):
+        if self.unit == 'B':
+            return self.val
+        elif self.unit == 'K':
+            return self.val * 1024
+        elif self.unit == 'M':
+            return self.val * 1048576
+        elif self.unit == 'G':
+            return self.val * 1073741824
+        elif self.unit == 'T':
+            return self.val * 1099511627776
+
+    def __cmp__(self, other):
+        return cmp(self.bytes, other.bytes)
