+++ src/toil/batchSystems/slurm.py
@@ -23,42 +23,11 @@ from Queue import Queue, Empty
 from threading import Thread
 
 from toil.batchSystems.abstractBatchSystem import AbstractBatchSystem
+from toil.batchSystems import MemoryString
 
 logger = logging.getLogger(__name__)
 
 sleepSeconds = 1
-
-
-class MemoryString:
-    def __init__(self, string):
-        if string[-1] == 'K' or string[-1] == 'M' or string[-1] == 'G' or string[-1] == 'T':
-            self.unit = string[-1]
-            self.val = float(string[:-1])
-        else:
-            self.unit = 'B'
-            self.val = float(string)
-        self.bytes = self.byteVal()
-
-    def __str__(self):
-        if self.unit != 'B':
-            return str(self.val) + self.unit
-        else:
-            return str(self.val)
-
-    def byteVal(self):
-        if self.unit == 'B':
-            return self.val
-        elif self.unit == 'K':
-            return self.val * 1024
-        elif self.unit == 'M':
-            return self.val * 1048576
-        elif self.unit == 'G':
-            return self.val * 1073741824
-        elif self.unit == 'T':
-            return self.val * 1099511627776
-
-    def __cmp__(self, other):
-        return cmp(self.bytes, other.bytes)
 
 
 class Worker(Thread):
